pyproject.toml
[tool.poetry]
name = "mlops-pricing-optimization"
version = "0.1.0"
description = "Systeme MLOps d'optimisation des prix basee sur l'estimation de la demande"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Data processing
pandas = "^2.2.0"
numpy = "^1.26.0"
pyarrow = "^15.0.0"
# Machine Learning
scikit-learn = "^1.4.0"
lightgbm = "^4.3.0"
optuna = "^3.5.0"
# MLOps
mlflow = "^2.10.0"
dvc = "^3.42.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
# Optimization
scipy = "^1.12.0"
# API
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
# UI
streamlit = "^1.31.0"
plotly = "^5.18.0"
# Utilities
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
joblib = "^1.3.0"
tqdm = "^4.66.0"
kaggle = "^1.6.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
# Quality
ruff = "^0.2.0"
mypy = "^1.8.0"
# Type stubs
pandas-stubs = "^2.1.0"
types-tqdm = "^4.66.0"

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "B905",   # zip without strict (Python 3.10+ only)
    "SIM108", # ternary operator (prefer explicit if/else for readability)
    "ARG001", # unused function argument (often intentional for interface)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
